Rewrite 'has_threadmarks' into 'threadmark_count', use slightly advanced SQL to update/insert in the single function.

+++ upload/library/Sidane/Threadmarks/Install.php
@@ -4,10 +4,6 @@ class Sidane_Threadmarks_Install
 {
   public static function install($existingAddOn, $addOnData)
   {
-    if ($existingAddOn['version_id'] > 1) {
-      return;
-    }
-
     $db = XenForo_Application::get('db');
 
     $db->query("
@@ -16,23 +12,33 @@ class Sidane_Threadmarks_Install
         thread_id INT UNSIGNED NOT NULL,
         post_id INT UNSIGNED NOT NULL,
         label VARCHAR(100) NOT NULL,
-        KEY thread_id (thread_id)
+        UNIQUE KEY `thread_post_id` (`thread_id`,`post_id`)
       ) ENGINE = InnoDB CHARACTER SET utf8 COLLATE utf8_general_ci
     ");
 
-    try
+    if ($existingAddOn['version_id'] <= 1) 
     {
-      $db->query("ALTER TABLE xf_thread ADD COLUMN has_threadmarks INT UNSIGNED DEFAULT 0 NOT NULL AFTER prefix_id");
+        try
+        {
+          $db->query("ALTER TABLE xf_thread DROP COLUMN has_threadmarks;");
+        }
+        catch (Zend_Db_Exception $e) {}
     }
-    catch (Zend_Db_Exception $e) {}
+    
+    if ($existingAddOn['version_id'] < 2) 
+    {    
+        try
+        {
+          $db->query("ALTER TABLE xf_thread ADD COLUMN threadmark_count INT UNSIGNED DEFAULT 0 NOT NULL");
+        }
+        catch (Zend_Db_Exception $e) {} 
+    }    
   }
 
   public static function uninstall()
   {
     $db = XenForo_Application::get('db');
-    $db->query("ALTER TABLE xf_thread DROP COLUMN has_threadmarks");
+    $db->query("ALTER TABLE xf_thread DROP COLUMN threadmark_count");
     $db->query("DROP TABLE threadmarks");
   }
 }
-
-?>
